refactor: Refatoração do serviço de contabilização.
