+++ dbpautas-back-end/src/main/java/com/db/dbpautasbackend/controller/PautaController.java
@@ -5,9 +5,6 @@ import com.db.dbpautasbackend.dto.PautaFinalizadaDTO;
 import com.db.dbpautasbackend.dto.RegistrarPautaDTO;
 import com.db.dbpautasbackend.enums.Categoria;
 import com.db.dbpautasbackend.enums.Voto;
-import com.db.dbpautasbackend.mapper.PautaMapper;
-import com.db.dbpautasbackend.model.Pauta;
-import com.db.dbpautasbackend.service.ContabilizacaoService;
 import com.db.dbpautasbackend.service.PautaService;
 import io.swagger.v3.oas.annotations.Operation;
 import jakarta.validation.Valid;
