+++ dbpautas-back-end/src/main/java/com/db/dbpautasbackend/mapper/PautaMapper.java
@@ -4,7 +4,7 @@ import com.db.dbpautasbackend.dto.PautaEmAndamentoDTO;
 import com.db.dbpautasbackend.dto.PautaFinalizadaDTO;
 import com.db.dbpautasbackend.dto.RegistrarPautaDTO;
 import com.db.dbpautasbackend.model.Pauta;
-import com.db.dbpautasbackend.service.ContabilizacaoService;
+import com.db.dbpautasbackend.service.PautaService;
 
 import java.util.List;
 
@@ -30,7 +30,7 @@ public interface PautaMapper {
         ).toList();
     }
 
-    static List<PautaFinalizadaDTO> mapListOfPautaToListOfPautaFinalizadaDTO(List<Pauta> pautas, ContabilizacaoService contabilizacaoService){
+    static List<PautaFinalizadaDTO> mapListOfPautaToListOfPautaFinalizadaDTO(List<Pauta> pautas, PautaService pautaService){
         return pautas.stream().map(pauta -> PautaFinalizadaDTO.builder().id(pauta.getId())
                         .titulo(pauta.getTitulo())
                         .resumo(pauta.getResumo())
@@ -38,7 +38,7 @@ public interface PautaMapper {
                         .categoria(pauta.getCategoria())
                         .votosSim(pauta.getVotosSim())
                         .votosNao(pauta.getVotosNao())
-                        .decisao(contabilizacaoService.contabilizar(pauta))
+                        .decisao(pautaService.contabilizar(pauta))
                         .build()
         ).toList();
     }
