--- dbpautas-back-end/src/main/java/com/db/dbpautasbackend/service/ContabilizacaoService.java
@@ -1,8 +1,0 @@
-package com.db.dbpautasbackend.service;
-
-import com.db.dbpautasbackend.enums.Decisao;
-import com.db.dbpautasbackend.model.Pauta;
-
-public interface ContabilizacaoService {
-    Decisao contabilizar(Pauta pauta);
-}
